requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1

chromadb==0.5.5

openai==1.57.0
httpx=0.27.2

pypdf==4.3.1
python-pptx==0.6.23
Pillow==10.4.0

# OCR
pytesseract==0.3.10
